Update and rename test/com/promo/model/ItemTest.java to test/com/promo/service/CartServiceTest.java
